+++ backend/canisters/notifications_index/api/src/lib.rs
@@ -16,12 +16,16 @@ pub enum NotificationsIndexEvent {
     AllSubscriptionsRemoved(UserId),
     UserBlocked(UserId, UserId),
     UserUnblocked(UserId, UserId),
+    BotEndpointUpdated(UserId, String),
+    BotRemoved(UserId),
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub enum UserIndexEvent {
     UserBlocked(UserId, UserId),
     UserUnblocked(UserId, UserId),
+    BotEndpointUpdated(UserId, String),
+    BotRemoved(UserId),
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]